feat(client): add Long256 support (#7)
* add: Long256Column() function

The support for long type has been addded. The implementation is done, but it doesn't seem to be working as expected.

* implemented long256Colums() using string as parameter

The string method is a temporary fix, it supports data entry in long256 until 64bit. After that it thows an error. Examine it in examples/long/long.go

* feat: support for long256 added

func Long256Column() added along with unit test.

* add: Integration test for long256

* typos: fixed minor typos and added suggestions

* fix: expanded defination